neural/requirements.txt
neat-python==0.92
pygame==2.5.1
